test progress
Seems like the code for just views is running. But for the forms, the post request testing logic isn't getting tested all the way. I think most will be pretty straightforward, but I'm not sure about the edit, and how to interact with the database from within the views function.